requirements.txt
streamlit==1.45.0
opencv-python-headless==4.10.0.84
tensorflow==2.19.0
pillow==11.2.1
numpy==1.26.4

